requirements.txt
Django>=3.2,<4.0
gunicorn>=20.1
mysqlclient>=2.0
whitenoise>=5.2.0
django-heroku>=0.3.1
djangorestframework>=3.12.4
djangorestframework-simplejwt>=4.8.0
python-dotenv>=0.19.0
requests>=2.25.1
pytest>=6.2.5
pytest-django>=4.4.0